requirements.txt
# ============================================================================
# DEPENDENCIAS DEL PROYECTO DE DECODIFICACIÓN DE IMÁGENES MENTALES
# ============================================================================
# Basado en Koide-Majima et al. (2024)
# "Mental image reconstruction from human brain activity"
#
# Instalación:
#   pip install -r requirements.txt
#
# Nota: Para GPU, instalar PyTorch con CUDA:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# ============================================================================

# ============================================================================
# CORE: Deep Learning Frameworks
# ============================================================================
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0

# ============================================================================
# MODELOS: Arquitecturas específicas
# ============================================================================
# CLIP - Alineación semántica
ftfy>=6.0.0
regex>=2021.0.0
tqdm>=4.62.0

# VQGAN - Generación de imágenes
omegaconf>=2.1.0
taming-transformers-rom1504  # Versión compatible

# VGG19 viene incluido en torchvision.models

# ============================================================================
# PROCESAMIENTO DE DATOS
# ============================================================================
Pillow>=8.3.0           # Manejo de imágenes
scipy>=1.7.0            # Lectura de archivos .mat
pickle5>=0.0.11         # Lectura de archivos .pkl (Python <3.8)

# ============================================================================
# UTILIDADES
# ============================================================================
matplotlib>=3.4.0       # Visualización (opcional)
pandas>=1.3.0           # Análisis de datos (opcional)
seaborn>=0.11.0         # Visualización avanzada (opcional)

# ============================================================================
# LOGGING Y PROGRESO
# ============================================================================
coloredlogs>=15.0       # Logs con colores (opcional)
rich>=10.0.0            # Output enriquecido (opcional)

# ============================================================================
# CONFIGURACIÓN RECOMENDADA POR HARDWARE
# ============================================================================
# GPU (NVIDIA):
#   torch==1.13.0+cu117
#   torchvision==0.14.0+cu117
#   Instalar con: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu117
#
# CPU (sin GPU):
#   torch==1.13.0
#   torchvision==0.14.0
#   Instalar con: pip install torch torchvision
#
# Verificar instalación:
#   python -c "import torch; print(f'PyTorch: {torch.__version__}'); print(f'CUDA: {torch.cuda.is_available()}')"
# ============================================================================

# ============================================================================
# DEPENDENCIAS OPCIONALES (para desarrollo)
# ============================================================================
# jupyter>=1.0.0          # Notebooks interactivos
# tensorboard>=2.6.0      # Monitoreo de entrenamiento
# opencv-python>=4.5.0    # Procesamiento de video/imágenes avanzado
# scikit-learn>=0.24.0    # Métricas de evaluación
